+++ packages/terra-date-picker/CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## Unreleased
 
+## 4.97.0 - (January 25, 2024)
+
+* Changed
+  * Updated metadata as argument for onBlur callback.
+
 ## 4.96.0 - (January 22, 2024)
 
 * Changed